+++ docs/changelog.md
@@ -4,9 +4,19 @@ _This project uses semantic versioning_
 
 ## UNRELEASED
 
-- Upgrade [egglog](https://github.com/egraphs-good/egglog/compare/0113af1d6476b75d4319591cc3d675f96a71cdc5...fb4a9f114f9bb93154d6eff0dbab079b5cb4ebb6)
-  - Adds `UnstableCombinedRulset` to commands
+## New Feaatures
+
+- Upgrade [egglog](https://github.com/egraphs-good/egglog/compare/0113af1d6476b75d4319591cc3d675f96a71cdc5...fb4a9f114f9bb93154d6eff0dbab079b5cb4ebb6) ([#143](https://github.com/egraphs-good/egglog-python/pull/143))
+  - Adds `bindings.UnstableCombinedRulset` to commands
   - Adds `UnstableFn` sort
+- Adds support for first class functions as values using Python's built in `Callable` syntax and `partial`.
+- Adds way to combine ruleset with `r1 | r2` syntax or the experimental `unstable_combine_rulesets(*rs, name=None)` function.
+
+## Minor improvements
+
+- Fixes a bug where you could not write binary dunder methods (like `__add__`) that didn't have symetric arguments
+- Use function name as ruleset name by default when creating ruleset from function
+- Adds ability to refer to methods and property off of classes instead of only off of instances (i.e. `Math.__add__(x, y)`)
 
 ## 7.0.0 (2024-04-27)